requirements.txt
sentry-sdk
boto3
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
httpx==0.27.2
PyYAML==6.0.2
Jinja2==3.1.4
psycopg2-binary
# optional later:

feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.32.3
APScheduler==3.10.4
GitPython==3.1.43
